+++ extensions-core/multi-stage-query/src/test/java/org/apache/druid/msq/indexing/MSQControllerTaskTest.java
@@ -59,6 +59,7 @@ public class MSQControllerTaskTest
           Granularities.DAY,
           null,
           INTERVALS,
+          null,
           null
       ))
       .query(new Druids.ScanQueryBuilder()
