+++ extensions-core/multi-stage-query/src/test/java/org/apache/druid/msq/sql/resources/SqlStatementResourceTest.java
@@ -200,6 +200,7 @@ public class SqlStatementResourceTest extends MSQTestBase
                  Granularities.DAY,
                  null,
                  null,
+                 null,
                  null
              ))
              .tuningConfig(
