+++ extensions-core/multi-stage-query/src/test/java/org/apache/druid/msq/test/MSQTestDelegateDataSegmentPusher.java
@@ -40,7 +40,7 @@ public class MSQTestDelegateDataSegmentPusher implements DataSegmentPusher
       MSQTestSegmentManager segmentManager
   )
   {
-    delegate = dataSegmentPusher;
+    this.delegate = dataSegmentPusher;
     this.segmentManager = segmentManager;
   }
