+++ extensions-core/multi-stage-query/src/test/java/org/apache/druid/msq/util/SqlStatementResourceHelperTest.java
@@ -376,6 +376,7 @@ public class SqlStatementResourceHelperTest
             Granularities.DAY,
             null,
             null,
+            null,
             null
         )
     );
